Add property observer to tokens for saving
The documentation recommends we the tokens to `saveToUserDefaults()` in `applicationWillTerminate`. This is an unsafe option because that method is not *guaranteed* to be called in instances like low memory failures. It is safest to add them to user defaults when the token changes and isn't a huge performance downgrade since the `synchronize()` call was removed.